Change GuestPricePerNight type in GetAllPropertiesDto

+++ server/PropertyPro.Infrastructure/Dtos/Property/GetAllPropertiesDto.cs
@@ -23,7 +23,7 @@ namespace PropertyPro.Infrastructure.Dtos.Property
 
         public string Country { get; set; } = null!;
 
-        public double GuestPricePerNight { get; set; }
+        public decimal GuestPricePerNight { get; set; }
 
         public int MaxGuestsCount { get; set; }
